Move PokemonListJson to shared package
