+++ src/main/java/spyra/lukasz/pokerestapi/shared/PokemonListJson.java
@@ -1,4 +1,4 @@
-package spyra.lukasz.pokerestapi.consume;
+package spyra.lukasz.pokerestapi.shared;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import org.slf4j.Logger;
@@ -8,17 +8,18 @@ import java.util.List;
 import java.util.stream.Collectors;
 
 @JsonIgnoreProperties(ignoreUnknown = true)
-record PokemonListJson(int count, List<ResourceResult> results) {
+public record PokemonListJson(int count, List<ResourceResult> results) {
     private static final Logger log = LoggerFactory.getLogger(PokemonListJson.class);
 
     private record ResourceResult(String name, String url) {
     }
 
-    List<String> resourceUrls() {
+    public List<String> resourceUrls() {
+        log.debug("Extracting list of resources urls");
         return results.stream().map(ResourceResult::url).collect(Collectors.toList());
     }
 
-    PokemonListJson {
+    public PokemonListJson {
         log.debug("Create record instance for Pokemon list");
     }
 }
