+++ src/main/resources/settings.properties
@@ -1,3 +1,3 @@
 #API initial endpoint
-api.url=https://pokeapi.co/api/v2/pokemon?limit=10&offset=0
+api.root.url=https://pokeapi.co/api/v2/pokemon?limit=10&offset=0
