+++ src/test/integration/java/spyra/lukasz/pokerestapi/consume/PokemonListJsonParseTest.java
@@ -5,6 +5,7 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.client.RestOperations;
 import org.springframework.web.client.RestTemplate;
+import spyra.lukasz.pokerestapi.shared.PokemonListJson;
 
 import java.util.Objects;
